pyproject.toml
[project]
name = "biomonitor-capstone"
version = "0.1.0"
description = "Global Mammal Biodiversity Pipeline: End-to-end data platform (API -> GCS -> BigQuery -> dbt)"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "apache-airflow>=3.2.0",
    "dbt-bigquery>=1.3.3",
    "dlt[bigquery]",
    "gcsfs",
    "google-cloud-bigquery-storage",
    "google-cloud-storage>=3.1.1",
    "pandas",
    "pyarrow",
    "python-dotenv>=1.2.2",
]
